+++ src/toolchain/mod.rs
@@ -229,7 +229,7 @@ pub async fn install(args: InstallOpts, install_mode: InstallMode) -> Result<()>
     let host_triple = get_host_triple(args.default_host)?;
     let xtensa_rust_version = if let Some(toolchain_version) = &args.toolchain_version {
         if !args.skip_version_parse {
-            XtensaRust::parse_version(toolchain_version)?
+            XtensaRust::find_latest_version_on_github(toolchain_version)?
         } else {
             toolchain_version.clone()
         }